chore: expose mockNostoRecs
