+++ test/components/NostoSection.spec.ts
@@ -1,9 +1,9 @@
 import { describe, it, beforeEach, expect, vi, Mock } from "vitest"
-import { mockNostojs } from "@nosto/nosto-js/testing"
 import { NostoSection } from "@/components/NostoSection/NostoSection"
-import { JSONProduct, RequestBuilder } from "@nosto/nosto-js/client"
+import { RequestBuilder } from "@nosto/nosto-js/client"
 import { addHandlers } from "../msw.setup"
 import { http, HttpResponse } from "msw"
+import { mockNostoRecs } from "../mockNostoRecs"
 
 // TODO replace with vi fake timers
 async function flushBatches() {
@@ -22,7 +22,7 @@ describe("NostoSection", () => {
 
   it("renders section markup from product handles and attributes product clicks", async () => {
     const products = [{ handle: "product-a" }, { handle: "product-b" }]
-    const { attributeProductClicksInCampaign, load, mockBuilder } = mockNosto({ products })
+    const { attributeProductClicksInCampaign, load, mockBuilder } = mockNostoRecs("placement1", { products })
 
     const sectionHTML = `<div class=\"wrapper\"><div class=\"inner\">Rendered Section</div></div>`
     addHandlers(
@@ -50,7 +50,7 @@ describe("NostoSection", () => {
   })
 
   it("throws when section fetch fails", async () => {
-    mockNosto({ products: [{ handle: "x" }] })
+    mockNostoRecs("placement1", { products: [{ handle: "x" }] })
 
     addHandlers(http.get("/search", () => HttpResponse.text("Error", { status: 500 })))
 
@@ -62,33 +62,3 @@ describe("NostoSection", () => {
     expect(el.hasAttribute("loading")).toBe(false)
   })
 })
-
-function mockNosto(result: { products: Partial<JSONProduct>[] }) {
-  const load = vi.fn().mockResolvedValue({
-    recommendations: {
-      placement1: result
-    }
-  })
-
-  const mockBuilder: Partial<RequestBuilder> = {
-    disableCampaignInjection: () => mockBuilder as RequestBuilder,
-    setElements: vi.fn(() => mockBuilder as RequestBuilder),
-    setResponseMode: vi.fn(() => mockBuilder as RequestBuilder),
-    setProducts: vi.fn(() => mockBuilder as RequestBuilder),
-    load
-  }
-
-  const attributeProductClicksInCampaign = vi.fn()
-
-  const api = {
-    createRecommendationRequest: () => mockBuilder as RequestBuilder,
-    attributeProductClicksInCampaign
-  }
-  mockNostojs(api)
-
-  return {
-    load,
-    mockBuilder,
-    attributeProductClicksInCampaign
-  }
-}
